refactor(api): updated import of moved files
