+++ api/src/certification/enrolment/application/attendance-sheet-route.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
 
-import { authorization } from '../../../../lib/application/preHandlers/authorization.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
+import { authorization } from '../../shared/application/pre-handlers/authorization.js';
 import { attendanceSheetController } from './attendance-sheet-controller.js';
 
 const register = async function (server) {
