+++ api/src/certification/session-management/application/certification-report-route.js
@@ -1,8 +1,8 @@
 import Joi from 'joi';
 
-import { authorization } from '../../../../lib/application/preHandlers/authorization.js';
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
+import { authorization } from '../../shared/application/pre-handlers/authorization.js';
 import { certificationReportController } from './certification-report-controller.js';
 
 const register = async function (server) {
