+++ api/tests/certification/enrolment/unit/application/session-route_test.js
@@ -1,6 +1,6 @@
-import { authorization } from '../../../../../lib/application/preHandlers/authorization.js';
 import { sessionController } from '../../../../../src/certification/enrolment/application/session-controller.js';
 import * as moduleUnderTest from '../../../../../src/certification/enrolment/application/session-route.js';
+import { authorization } from '../../../../../src/certification/shared/application/pre-handlers/authorization.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
